Add new permissions for read/write social stream.
This is a manual merge of a change going in to ICS-FactoryROM.

These permissions are needed to separate the (potentially invasive)
access to the user's social stream from the existing read/write
contacts permission.

Per discussion with Android release team, we are also hiding the
stream item API until we figure out a better way to guard the data.

Bug 5406886

Change-Id: I8339d743c3ebe8923c7ee47f2900444efcf82a52