Visual glitches when starting extracted text
Partially addresses Bug 5547923

In landscape, tapping on the text
1. starts extracted text mode
2. displays the cursor handle or the selection popup window.

As a result, some ghost effects and race conditions create an
unpleasing visual experience.

Fixed this by not doing 2. in case extracted mode will start.

The drawback of this quicl fix is that the user will have to
tap again to get the handle/suggestion. That can be fixed later
if needed.

Change-Id: I10e1d8399bb35e5b2cd5cba1295f7d29d051cae0